agents/oleg/requirements.txt
# Telegram Bot Framework
aiogram==3.15.0
aiogram_calendar==0.6.0

# AI Providers
httpx==0.27.0

# Database
psycopg2-binary==2.9.9

# Password Hashing
bcrypt==4.2.0

# Scheduling
APScheduler==3.10.4
pytz==2024.1

# Utilities
python-dotenv==1.0.0
